Moved dues & currency logic to separate package
