+++ src/main/java/com/pretz/parkingmanager/calculator/currency/CurrencyConverter.java
@@ -1,7 +1,8 @@
-package com.pretz.parkingmanager.service;
+package com.pretz.parkingmanager.calculator.currency;
 
 import java.math.BigDecimal;
 
 public interface CurrencyConverter {
+
     BigDecimal getCurrencyMultiplier();
 }
