+++ src/main/java/com/pretz/parkingmanager/service/DuesCheckService.java
@@ -1,5 +1,7 @@
 package com.pretz.parkingmanager.service;
 
+import com.pretz.parkingmanager.calculator.currency.CurrencyConverter;
+import com.pretz.parkingmanager.calculator.DuesCalculator;
 import com.pretz.parkingmanager.domain.ParkingSession;
 import com.pretz.parkingmanager.dto.DuesRequestDTO;
 import com.pretz.parkingmanager.dto.DuesResponseDTO;
