--- src/main/java/com/pretz/parkingmanager/service/GenericCurrencyConverter.java
@@ -1,14 +1,0 @@
-package com.pretz.parkingmanager.service;
-
-import com.pretz.parkingmanager.exception.UnknownCurrencyException;
-import org.springframework.stereotype.Component;
-
-import java.math.BigDecimal;
-
-@Component
-public class GenericCurrencyConverter implements CurrencyConverter {
-    @Override
-    public BigDecimal getCurrencyMultiplier() {
-        throw new UnknownCurrencyException();
-    }
-}
